kv-store/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
kubernetes==31.0.0
pydantic==2.9.2